+++ libraries/ui-library/stencil.config.ts
@@ -52,9 +52,27 @@ export const config: Config = {
       type: 'dist',
       esmLoaderPath: '../loader',
       copy: [
-        { src: '../../../node_modules/@fontsource/material-icons/files/*', dest: './files', warn: true },
-        { src: '../../../node_modules/@fontsource/material-icons-outlined/files/*', dest: './files', warn: true },
-        { src: '../../../node_modules/@fontsource/noto-sans/files/*', dest: './files', warn: true },
+        { src: '../../../node_modules/@fontsource/material-icons/files/**/*', dest: './files', warn: true },
+        { src: '../../../node_modules/@fontsource/material-icons-outlined/files/**/*', dest: './files', warn: true },
+        { src: '../../../node_modules/@fontsource/noto-sans/files/**/*', dest: './files', warn: true },
+
+        // NEW: Material Symbols (use /**/* for subdirectories)
+        {
+          src: '../../../node_modules/@fontsource-variable/material-symbols-outlined/files/**/*',
+          dest: './files',
+          warn: true,
+        },
+        {
+          src: '../../../node_modules/@fontsource-variable/material-symbols-rounded/files/**/*',
+          dest: './files',
+          warn: true,
+        },
+        {
+          src: '../../../node_modules/@fontsource-variable/material-symbols-sharp/files/**/*',
+          dest: './files',
+          warn: true,
+        },
+
         { src: './assets/**/*', dest: '.', warn: true, keepDirStructure: true },
       ],
     },
@@ -83,9 +101,31 @@ export const config: Config = {
       type: 'www',
       serviceWorker: null, // disable service workers
       copy: [
-        { src: '../../../node_modules/@fontsource/material-icons/files/*', dest: 'build/files', warn: true },
-        { src: '../../../node_modules/@fontsource/material-icons-outlined/files/*', dest: 'build/files', warn: true },
-        { src: '../../../node_modules/@fontsource/noto-sans/files/*', dest: 'build/files', warn: true },
+        { src: '../../../node_modules/@fontsource/material-icons/files/**/*', dest: 'build/files', warn: true },
+        {
+          src: '../../../node_modules/@fontsource/material-icons-outlined/files/**/*',
+          dest: 'build/files',
+          warn: true,
+        },
+        { src: '../../../node_modules/@fontsource/noto-sans/files/**/*', dest: 'build/files', warn: true },
+
+        // NEW: Material Symbols (use /**/* for subdirectories)
+        {
+          src: '../../../node_modules/@fontsource-variable/material-symbols-outlined/files/**/*',
+          dest: 'build/files',
+          warn: true,
+        },
+        {
+          src: '../../../node_modules/@fontsource-variable/material-symbols-rounded/files/**/*',
+          dest: 'build/files',
+          warn: true,
+        },
+        {
+          src: '../../../node_modules/@fontsource-variable/material-symbols-sharp/files/**/*',
+          dest: 'build/files',
+          warn: true,
+        },
+
         { src: '**/*.html' },
       ],
     },